python/salesforce_connector.py
"""
Salesforce Integration Module

Handles authentication and API calls to Salesforce
"""

import os
from typing import Optional, Dict, Any
from dotenv import load_dotenv

load_dotenv()


class SalesforceConnector:
    """Connect to Salesforce and execute operations"""
    
    def __init__(self, instance_url: Optional[str] = None):
        """Initialize Salesforce connector"""
        self.instance_url = instance_url or os.getenv("SALESFORCE_INSTANCE_URL")
        self.client_id = os.getenv("SALESFORCE_CLIENT_ID")
        self.client_secret = os.getenv("SALESFORCE_CLIENT_SECRET")
    
    def authenticate(self) -> Dict[str, Any]:
        """Authenticate with Salesforce"""
        # Placeholder for authentication logic
        return {
            "status": "authenticated",
            "instance": self.instance_url
        }
    
    def query(self, soql: str) -> Dict[str, Any]:
        """Execute SOQL query"""
        # Placeholder for query execution
        return {
            "totalSize": 0,
            "records": []
        }
