python/requirements.txt
salesforce-sdk==1.0.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
aiohttp==3.9.1
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
pylint==3.0.3
